+++ src/main/java/org/betonquest/betonquest/compatibility/mmogroup/mmoitems/take/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.event.Event Event} implementation for MMOItems take events.
+ */
+package org.betonquest.betonquest.compatibility.mmogroup.mmoitems.take;
