+++ src/main/java/org/betonquest/betonquest/quest/registry/CoreQuestTypes.java
@@ -49,7 +49,6 @@ import org.betonquest.betonquest.conditions.TimeCondition;
 import org.betonquest.betonquest.conditions.VariableCondition;
 import org.betonquest.betonquest.conditions.WeatherCondition;
 import org.betonquest.betonquest.conditions.WorldCondition;
-import org.betonquest.betonquest.events.TakeEvent;
 import org.betonquest.betonquest.objectives.ActionObjective;
 import org.betonquest.betonquest.objectives.ArrowShootObjective;
 import org.betonquest.betonquest.objectives.BlockObjective;
@@ -138,6 +137,7 @@ import org.betonquest.betonquest.quest.event.spawn.SpawnMobEventFactory;
 import org.betonquest.betonquest.quest.event.stage.StageEventFactory;
 import org.betonquest.betonquest.quest.event.tag.TagGlobalEventFactory;
 import org.betonquest.betonquest.quest.event.tag.TagPlayerEventFactory;
+import org.betonquest.betonquest.quest.event.take.TakeEventFactory;
 import org.betonquest.betonquest.quest.event.teleport.TeleportEventFactory;
 import org.betonquest.betonquest.quest.event.time.TimeEventFactory;
 import org.betonquest.betonquest.quest.event.variable.VariableEventFactory;
@@ -325,7 +325,7 @@ public class CoreQuestTypes {
         eventTypes.register("stage", new StageEventFactory(betonQuest));
         eventTypes.register("sudo", new SudoEventFactory(loggerFactory, data));
         eventTypes.registerCombined("tag", new TagPlayerEventFactory(betonQuest, betonQuest.getSaver()));
-        eventTypes.register("take", TakeEvent.class);
+        eventTypes.register("take", new TakeEventFactory(loggerFactory));
         eventTypes.register("teleport", new TeleportEventFactory(loggerFactory, data));
         eventTypes.registerCombined("time", new TimeEventFactory(server, data, betonQuest.getVariableProcessor()));
         eventTypes.register("variable", new VariableEventFactory(betonQuest, betonQuest.getVariableProcessor()));
